+++ packages/template-retail-react-app/app/utils/product-utils.js
@@ -5,6 +5,8 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import {productUrlBuilder, rebuildPathWithParams} from '@salesforce/retail-react-app/app/utils/url'
+
 /**
  * Get the human-friendly version of the variation values that users have selected.
  * Useful for displaying these values in the UI.
@@ -89,6 +91,75 @@ export const getPriceData = (product, opts = {}) => {
         tieredPrice: closestTieredPrice?.price,
         maxPrice: product?.priceMax || maxTieredPrice
     }
+}
+
+/**
+ * Filters the provided product `ImageGroups` array with the optional filters.
+ *
+ * @param {ImageGroup[]} imageGroups
+ * @param {object} filters
+ * @param {string} filters.viewType
+ * @param {object} filters.variationValues
+ *
+ * @returns {ImageGroup[]} filteredImageGroups
+ */
+export const filterImageGroups = (imageGroups = [], filters) => {
+    if (!filters) {
+        throw new Error(`Missing required "filters" argument`)
+    }
+
+    const {viewType, variationValues = {}} = filters
+    // NOTE: For future feature enhancement compatibility we use an object for the `variationValues` as
+    // we might want to filter on more than the one attribute id/value pair.
+    const [selectedAttributeId, selectedAttributeValue] = Object.entries(variationValues)[0] || []
+
+    // Filters
+    const typeFilter = viewType ? (group) => group.viewType === viewType : () => true
+    const attributeFilter =
+        !!selectedAttributeId && !!selectedAttributeValue
+            ? ({variationAttributes = []}) =>
+                  variationAttributes.some(
+                      ({id, values}) =>
+                          id === selectedAttributeId &&
+                          !!values.find(({value}) => value === selectedAttributeValue)
+                  )
+            : () => true
+
+    return imageGroups?.filter(typeFilter)?.filter(attributeFilter)
+}
+
+/**
+ * Provided a product this function will return the variation attibutes decorated with
+ * `href` and `swatch` image for the given attribute values. This allows easier access
+ * when creating components that commonly use this information.
+ *
+ * @param {Product} product
+ * @param {object} opts
+ * @param {string} opts.swatchViewType - The `viewTypeId` for the swatch image. Defaults to `swatch`.
+ *
+ * @returns {VariationAttributes[]} decoratedVariationAttributes
+ */
+export const getDecoratedVariationAttributes = (product, opts = {}) => {
+    const swatchViewType = opts.swatchViewType || 'swatch'
+
+    return product?.variationAttributes?.map((variationAttribute) => ({
+        ...variationAttribute,
+        values: variationAttribute.values.map((value) => {
+            const variationValues = {[variationAttribute.id]: value.value}
+
+            return {
+                ...value,
+                swatch: filterImageGroups(product.imageGroups, {
+                    viewType: swatchViewType,
+                    variationValues
+                })?.[0]?.images[0],
+                href: rebuildPathWithParams(
+                    productUrlBuilder({id: product.productId}),
+                    variationValues
+                )
+            }
+        })
+    }))
 }
 
 /**